feat: add consolidated plan selection page
